add some guards to the index add/remove

+++ db/migrate/20260117045120_enforce_membership_constraints.rb
@@ -5,9 +5,13 @@ class EnforceMembershipConstraints < ActiveRecord::Migration[8.0]
     # group_id already has an index to ensure it is not null
 
     # Remove the unique index on [:id, :group_id, :member_id]
-    remove_index :memberships, name: "index_memberships_on_id_and_group_id_and_member_id"
+    if index_exists?(:memberships, name: "index_memberships_on_id_and_group_id_and_member_id")
+      remove_index :memberships, name: "index_memberships_on_id_and_group_id_and_member_id"
+    end
 
     # Add a non-unique index on [:group_id, :member_type, :member_id]
-    add_index :memberships, [:group_id, :member_type, :member_id], name: "index_memberships_on_group_and_member"
+    unless index_exists?(:memberships, [:group_id, :member_type, :member_id], name: "index_memberships_on_group_and_member")
+      add_index :memberships, [:group_id, :member_type, :member_id], name: "index_memberships_on_group_and_member"
+    end
   end
 end
